moving css into main component noData css file
