+++ apps/modernization-ui/src/components/NoData/NoData.scss
@@ -1,0 +1,7 @@
+@import '../../settings.scss';
+
+.no-data {
+    color: $color-gray-30;
+    margin: 0;
+    font-style: italic;
+}
