+++ apps/modernization-ui/src/components/NoData/NoData.tsx
@@ -1,5 +1,7 @@
+import './NoData.scss';
+
 const NoData = ({ className = '' }: { className?: string }) => {
     return <span className={`no-data ${className}`}>No Data</span>;
 };
 
-export default NoData;
+export { NoData };
