+++ apps/modernization-ui/src/components/Table/style.scss
@@ -16,11 +16,6 @@ td {
     font-size: 16px !important;
 }
 
-.no-data {
-    color: #747474;
-    font-style: italic;
-}
-
 .show-length-text {
     color: #71767a;
 }
