+++ apps/modernization-ui/src/settings.scss
@@ -33,12 +33,6 @@ $theme-hero-image: '~uswds/src/img/hero.png';
     font-family: sans-serif !important;
 }
 
-.no-data {
-    color: $color-gray-30;
-    margin: 0;
-    font-style: italic;
-}
-
 td span {
     word-break: break-word;
 }
